Added saving the model

+++ main/__init__.py
@@ -4,20 +4,20 @@ import torchvision.transforms as transforms
 import matplotlib.pyplot as plt
 import numpy as np
 from network import Network
+import os.path as path
 
-trening_root = "D:\Data\Trainset\\"
+training_root = "D:\Data\Trainset\\"
 test_root = "D:\Data\Testset\\"
 classes = ['Inne', 'Makulatura', 'Plastik', 'Szklo']
-trainset = torchvision.datasets.ImageFolder(root=trening_root,
+model_path = "model.pt"
+train_set = torchvision.datasets.ImageFolder(root=training_root,
+                                             transform=transforms.ToTensor())
+training_loader = torch.utils.data.DataLoader(train_set, batch_size=8,
+                                              shuffle=True, num_workers=4)
+test_set = torchvision.datasets.ImageFolder(root=test_root,
                                             transform=transforms.ToTensor())
-trainingloader = torch.utils.data.DataLoader(trainset, batch_size=8,
-                                             shuffle=True, num_workers=4)
-testset = torchvision.datasets.ImageFolder(root=test_root,
-                                           transform=transforms.ToTensor())
-testloader = torch.utils.data.DataLoader(testset, batch_size=8,
-                                         shuffle=True, num_workers=4)
-
-
+test_loader = torch.utils.data.DataLoader(test_set, batch_size=8,
+                                          shuffle=True, num_workers=4)
 
 
 def main():
@@ -25,17 +25,25 @@ def main():
     # images, labels = dataiter.next()
     # print(' '.join('%5s' % classes[labels[j]] for j in range(8)))
     # imshow(torchvision.utils.make_grid(images))
-    net = Network.Net(lr=0.01, epochs=1, classes=classes,
-                      training_data_loader=trainingloader,
-                      test_data_loader=testloader)
-    net.to(net.device)
-    net._train()
-    plt.plot(net.loss_history)
-    plt.savefig('loss_history.png')
-    plt.show()
-    plt.plot(net.acc_history)
-    plt.savefig('acc_history.png')
-    plt.show()
+    net = None
+    if path.exists(model_path):
+        print("Model detected, loading...")
+        net = torch.load(model_path)
+    else:
+        net = Network.Net(lr=0.001, epochs=15, classes=classes,
+                          training_data_loader=training_loader,
+                          test_data_loader=test_loader)
+        net.to(net.device)
+        net._train()
+        plt.plot(net.loss_history)
+        plt.savefig('loss_history.png')
+        plt.show()
+        plt.plot(net.acc_history)
+        plt.savefig('acc_history.png')
+        plt.show()
+        net.eval()
+        torch.save(net, model_path)
+    print("Testing...")
     net._test()
